+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Module [AHAB](./docs/ahab/Ahab.md)
     - Added gate length as a parameter for operator `>` in note mode (#420)
     - Fixed crash on operator `<` when using whithout maximum value set (#425)
+- Module [MB](./docs/trial-and-error/Mb.md)
+    - Added fuzzy search to v1 browser (similar to the default module browser)
 
 ## 2.3.1